+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR3N6V48G17EDTDRSAKP5T6E"
+        "01KR448YCRCRVZ6DYE1B6CET2K"
     )
     r"""The unique identifier of the data source"""