+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR3N6V46XJQ6S1YP1VKQF8H9"
+        "01KR448YCP2AYRNVNV62YMW3VH"
     )
     r"""The unique identifier of the data source"""