+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-08T11:22:32.258Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-08T15:45:49.947Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")