+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR3N6V47RAY8ZYW7RDTWT781"
+        "01KR448YCQKJQ0K1T8A4JMKEZ3"
     )
     r"""The unique identifier of the data source"""