+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KR3N6V4AWCEREQ01FAGSWAG4"
+        "01KR448YCTPXSK4D8J6PAVMZKN"
     )
     r"""The unique identifier of the data source"""